Simpler object/pointer wrapping/unwrapping
Using `make-pointer` with the `finalizer` argument we avoid making
wrapped structs that are nothing else than a pain in the balls.

We are lucky because the `*_unref` functions match the finalizer
structure properly and we can use them directly.

This simplifies a lot the pointer usage, as we can send the pointers
created by `make-pointer` directly and we don't need to wrap/unwrap
anymore.